Added some SUPPORTED_UBUNTU_DISTROS options to better differentiate supported/unsupported ubuntus.
Added some vulnerability checks for CVE-2014-6277, CVE-2014-6278, CVE-2014-7186, CVE-2014-7187

Added in options to build/download an rpm for RH9

Extended build/download of RPM for RHEL4 to include RHEL3

Build for source extended from certain un-supported ubuntus and debians to all un-supported ubuntus and debians.  May not all work, but it is a last resort.